Feat: modular segregation and implementation of previous tasks using import and export
