+++ extra/pattern/diamond.js
@@ -1,4 +1,4 @@
-function diamond (n)
+export default function diamond (n)
 {
     let pattern='';
     for(let i=1;i<=n;i++)
@@ -33,4 +33,3 @@ function diamond (n)
     console.log(pattern);
 }
 
-diamond(5);
