+++ extra/pattern/index.js
@@ -1,0 +1,3 @@
+export {default as diamond} from './diamond';
+export {default as Equilateral_triangle} from './Equilateral.triangle';
+
