+++ extra/utils/helpers.js
@@ -1,0 +1,5 @@
+export function validateEmail(email) {
+    let regex = /^[\w-\.]+@(successive.tech)$/;
+
+    return regex.test(email);
+}
