+++ extra/utils/permissions.js
@@ -1,22 +1,16 @@
-const getUsers = "getUsers",
-  headTrainer = "head-trainer",
-  trainee = "trainee",
-  trainer = "trainer";
-const permissions = {
-  getUsers: {
-    all: [headTrainer],
-    read: [trainee, trainer],
-    write: [trainer],
-    delete: []
-  }
-};
+import { permissions } from "../constants";
 
-function hasPermission(moduleName, role, permissionType) {
+export function hasPermission(moduleName, role, permissionType) {
   if (
-    permissions[moduleName]['all'].includes(role) ||
+    permissions[moduleName]["all"].includes(role) ||
     permissions[moduleName][permissionType].includes(role)
-  )
-    return "True";
-  else return "False";
-}
-console.log(hasPermission("getUsers", "head-trainer", "delete"));
+  ){
+    console.log("True");
+    return ;
+  }  
+    else 
+{
+  console.log("False");
+  return;
+}}
+
