+++ extra/utils/validation.js
@@ -1,30 +1,16 @@
-const users = [
-  {
-    traineeEmail: "trainee1@successive.tech",
-    reviewerEmail: "reviewer1@succesfsive.tech"
-  },
-  {
-    traineeEmail: "trainee1@successive.tech",
-    reviewerEmail: "reviewer1@successive.tech"
-  },
-  {
-    traineeEmail: "trainee1@successive.tech",
-    reviewerEmail: "reviewer1@successive.tech"
-  }
-];
 
-function validateEmail(email) {
-  let regex = /^[\w-\.]+@(successive.tech)$/;
 
-  return regex.test(email);
-}
+import {validateEmail} from './helpers';
+
+
+export function validUser(users1) {
 
 let validuser = [];
 let invaliduser = [];
 let invalidcount = 0;
 let validcount = 0;
-function validUser(users) {
-  users.forEach(function(user) {
+
+  users1.forEach(function(user) {
     const { traineeEmail, reviewerEmail } = user;
     if (validateEmail(traineeEmail)) {
       validuser.push(traineeEmail);
@@ -47,4 +33,4 @@ function validUser(users) {
   console.log("\nThe count of invalid user is " + invalidcount);
   console.log("\nThe names of invalid users are :\n " + invaliduser);
 }
-validUser(users);
+
